Tests Example Local Script

+++ src/fake_csv.py
@@ -7,28 +7,6 @@ import sys
 
 filepath = ""
 
-if len(sys.argv) > 1:
-    filepath = sys.argv[1]
-
-if not isinstance(filepath, str) or not filepath:
-    filepath = "../dummydata.csv"
-    print(f"Filepath not included, defaulting to {filepath}\n")
-
-if os.access(filepath, os.F_OK) is True:
-    if os.access(filepath, os.W_OK) is False:
-        print(f"File {filepath} exists and is not writeable, Aborting")
-        sys.exit()
-    else:
-        overwrite = ""
-        while overwrite != "y" and overwrite != "n":
-            print(f"{overwrite=}")
-            overwrite = input(
-                f"File {filepath} already exists, overwrite? (y/n):"
-            ).lower()
-        if overwrite != "y":
-            print("Closing...")
-            sys.exit()
-
 
 class LanguageProvider(BaseProvider):
     def language(self):
@@ -45,22 +23,22 @@ ranks = ["Bronze", "Silver", "Gold", "Platinum", "Diamond", "Master"]
 rank_thresholds = [0, 25, 50, 75, 100, 125]
 
 
-def get_person_id():
+def _get_person_id():
     global person_id_count
     person_id_count += 1
     return person_id_count
 
 
-def get_street_address():
+def _get_street_address():
     return fake.street_address().replace("\n", " ").title()
 
 
-def get_phone_number():
+def _get_phone_number():
     number = random.randint(000000000, 999999999)
     return f"+44{number:09d}"
 
 
-def get_score():
+def _get_score():
     wins = random.randint(0, 50)
     losses = random.randint(0, 50)
     ties = random.randint(0, 50)
@@ -74,16 +52,16 @@ def get_score():
     return [played, wins, losses, ties, points, rank]
 
 
-def generate_data():
-    score = get_score()
+def _generate_data():
+    score = _get_score()
     return [
-        get_person_id(),
+        _get_person_id(),
         fake.first_name(),
         fake.last_name(),
-        get_street_address(),
+        _get_street_address(),
         fake.city(),
         fake.postcode(),
-        get_phone_number(),
+        _get_phone_number(),
         fake.date_between(start_date="-50y", end_date="-18y"),
         score[0],
         score[1],
@@ -95,28 +73,63 @@ def generate_data():
     ]
 
 
-with open(filepath, "w", newline="") as csvfile:
-    writer = csv.writer(csvfile)
-    writer.writerow(
-        [
-            "id",
-            "First Name",
-            "Last Name",
-            "Street Address",
-            "City",
-            "Postcode",
-            "Phone Number",
-            "Date of Birth",
-            "Played",
-            "Wins",
-            "Losses",
-            "Ties",
-            "Points",
-            "Rank",
-            "Language",
-        ]
-    )
-    file_size = 0
-    while file_size < 1:
-        file_size = os.stat(filepath).st_size / (1024 * 1024)
-        writer.writerow(generate_data())
+def fake_csv(filepath):
+
+    if not isinstance(filepath, str) or not filepath:
+        filepath = "../dummydata.csv"
+        print(f"Filepath not included, defaulting to {filepath}\n")
+
+    check = filepath.rsplit(".", 1)[1]
+    if check != "csv":
+        raise Exception("Target file is not .csv extension")
+
+    if os.access(filepath, os.F_OK) is True:
+        if os.access(filepath, os.W_OK) is False:
+            print(f"File {filepath} exists and is not writeable, Aborting")
+            sys.exit()
+        else:
+            overwrite = ""
+            while overwrite != "y" and overwrite != "n":
+                print(f"{overwrite=}")
+                overwrite = input(
+                    f"File {filepath} already exists, overwrite? (y/n):"
+                ).lower()
+            if overwrite != "y":
+                print("Closing...")
+                sys.exit()
+
+    with open(filepath, "w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow(
+            [
+                "id",
+                "First Name",
+                "Last Name",
+                "Street Address",
+                "City",
+                "Postcode",
+                "Phone Number",
+                "Date of Birth",
+                "Played",
+                "Wins",
+                "Losses",
+                "Ties",
+                "Points",
+                "Rank",
+                "Language",
+            ]
+        )
+        file_size = 0
+        while file_size < 1:
+            file_size = os.stat(filepath).st_size / (1024 * 1024)
+            writer.writerow(_generate_data())
+
+        return filepath
+
+
+if __name__ == "__main__":
+    if len(sys.argv) > 1:
+        filepath = sys.argv[1]
+
+    print(filepath)
+    fake_csv(filepath)
